wrote script to call function with product array, added close to final styling
