+++ products/index.html
@@ -9,7 +9,7 @@
     <link rel="stylesheet" href="../styles/reset.css">
     <!-- <link rel="stylesheet" href="../styles/style.css"> -->
     <link rel="stylesheet" href="./products.css">
-    <title>Document</title>
+    <title>Products</title>
 </head>
 
 <body>
@@ -25,8 +25,8 @@
             </h2>
 
             <content>
-                <ul>
-                    <li class="product">
+                <ul id="product-list">
+                    <li id="sample" class="product">
                         <img src="../assets/bat.png"/>
                         <div class="product-middle">
                             <div class="name">Baseball Bat</div>
@@ -45,7 +45,7 @@
         <div class="copy gutter-right">&copy; 2020 Alchemy Code Lab</div>
     </footer>
 
-    <script type="module" src="app.js"></script>
+    <script type="module" src="products.js"></script>
 </body>
 
 </html>
